+++ src/components/CoreConcept.jsx
@@ -1,0 +1,9 @@
+export default function CoreConcept({title, description, image}) {
+    return (
+      <li>
+        <img src={image} alt={title} />
+        <h3>{title}</h3>
+        <p>{description}</p>
+      </li>
+    );
+  }
